+++ src/Photos/utils.ts
@@ -1,0 +1,3 @@
+export const toUnixtime = (datetime: Date): number => {
+	return Math.floor(datetime.getTime() / 1000);
+}
