+++ src/types.ts
@@ -1,0 +1,17 @@
+export type Feature = {
+	type: "Feature"
+	geometry: {
+		type: string
+		coordinates: [number, number][]
+	}
+	properties: Record<string, string>
+}
+export type GeoJson = {
+	type: "FeatureCollection"
+	features: Feature[]
+}
+
+export interface FrickrAPISettings {
+	key: string // API Key
+	secret?: string // API Secret
+}
